update metadata types
